moved mongo and cassandra endpoints to Constants
