+++ src/ConfigContext.ts
@@ -1,6 +1,5 @@
-import { JunoEndpoints, MongoProxyEndpointsV2 } from "Common/Constants";
+import { CassandraProxyEndpoints, JunoEndpoints, MongoProxyEndpoints } from "Common/Constants";
 import {
-  CassandraProxyEndpoints,
   allowedAadEndpoints,
   allowedArcadiaEndpoints,
   allowedCassandraProxyEndpoints,
@@ -13,7 +12,7 @@ import {
   allowedMsalRedirectEndpoints,
   defaultAllowedArmEndpoints,
   defaultAllowedBackendEndpoints,
-  validateEndpoint
+  validateEndpoint,
 } from "Utils/EndpointUtils";
 
 export enum Platform {
@@ -89,7 +88,7 @@ let configContext: Readonly<ConfigContext> = {
   GITHUB_TEST_ENV_CLIENT_ID: "b63fc8cbf87fd3c6e2eb", // Registered OAuth app: https://github.com/organizations/AzureCosmosDBNotebooks/settings/applications/1777772
   JUNO_ENDPOINT: JunoEndpoints.Prod,
   BACKEND_ENDPOINT: "https://main.documentdb.ext.azure.com",
-  MONGO_PROXY_ENDPOINT: MongoProxyEndpointsV2.Prod,
+  MONGO_PROXY_ENDPOINT: MongoProxyEndpoints.Prod,
   NEW_MONGO_APIS: [
     // "resourcelist",
     // "createDocument",
